double changed to float

+++ ConvLayer.h
@@ -1,8 +1,8 @@
 #pragma once
 #include "Layer.h"
 
-typedef std::vector<Matrix<double>>  Kernel;
-typedef std::vector<Matrix<double>>  Images; // every matrix in vector is one Channel
+typedef std::vector<Matrix<float>>  Kernel;
+typedef std::vector<Matrix<float>>  Images; // every matrix in vector is one Channel
 
 
 struct Image_Dim
@@ -17,11 +17,11 @@ public:
 	ConvLayer(int Input_dim, int Output_dim, Image_Dim Image_dim, Image_Dim Kernel_Dim,ActivationFunction* Func_, MatrixInit init = MatrixInit::ZERO_INIT, WeightNormalization W_Init = WeightNormalization::None);
 	Tensor1D GetWeights();
 	Tensor1D GetBiases();
-	std::vector<Matrix<double>> ActivationPrime(Images& Z) override;
-	std::vector<Matrix<double>> ApplyActivation(Images& Z) override;
-    std::vector<Matrix<double>> Mul(Images& A) override;
-	std::vector<Matrix<double>> GetNablaWeight() override;
-	std::vector<Matrix<double>> GetNablaBias() override;
+	std::vector<Matrix<float>> ActivationPrime(Images& Z) override;
+	std::vector<Matrix<float>> ApplyActivation(Images& Z) override;
+    std::vector<Matrix<float>> Mul(Images& A) override;
+	std::vector<Matrix<float>> GetNablaWeight() override;
+	std::vector<Matrix<float>> GetNablaBias() override;
 	Tensor1D CalculateNablaWeight(const Tensor1D& Delta, const Tensor1D& Activation) override;
 	Tensor1D CalculateNablaBias(const Tensor1D& Delta, const Tensor1D& Activation) override;
 private:
@@ -29,7 +29,7 @@ private:
 	int Input_Dim;
 	ActivationFunction* Func = nullptr;
 	std::vector<Kernel> Kernels;
-	Matrix<double> Bias;
+	Matrix<float> Bias;
 	Image_Dim img_dim;
 };
