+++ CostFunction.cpp
@@ -1,14 +1,14 @@
 #include "CostFunction.h"
 #include <limits>
 
-Matrix<double> CostFunction::Function(const Matrix<double>& A, const Matrix<double>& Y)
+Matrix<float> CostFunction::Function(const Matrix<float>& A, const Matrix<float>& Y)
 {
-	return Matrix<double>();
+	return Matrix<float>();
 }
 
-Matrix<double> CostFunction::Function_Der(const Matrix<double>& A, const Matrix<double>& Y)
+Matrix<float> CostFunction::Function_Der(const Matrix<float>& A, const Matrix<float>& Y)
 {
-	return Matrix<double>();
+	return Matrix<float>();
 }
 
 CostFunction::~CostFunction()
@@ -16,12 +16,12 @@ CostFunction::~CostFunction()
 }
 
 
-Matrix<double> QuadraticCost::Function(const Matrix<double>& A, const Matrix<double>& Y)
+Matrix<float> QuadraticCost::Function(const Matrix<float>& A, const Matrix<float>& Y)
 {
 	return (A - Y) * (A - Y);
 }
 
-Matrix<double> QuadraticCost::Function_Der(const Matrix<double>& A, const Matrix<double>& Y)
+Matrix<float> QuadraticCost::Function_Der(const Matrix<float>& A, const Matrix<float>& Y)
 {
 	return (A - Y);
 }
